1.去掉了C api的支持，将来仅支持使用C++的接口进行处理 2.处理了宿主动态调用脚本函数的功能(回调) 3.减少用户程序对notstd库的依赖 4.修改一些测试代码来测试函数回调的功能 5.将substr由全局函数改为是string对象的成员函数 6.将项目迁移回vs2013 7.为在vs2013上消除大多数警告。 8.全平台全配置的批量生成的处理。 9.让对话框测试程序也能处理换行等字符 10.能够在gcc4.9.3（centos6.4）上编译通过 11.加入正规表达式判断功能 12.加入string.split函数 13.修改了文档 14.在windows上根据平台和配置，生成不同的库，便于最终的使用者使用

+++ core/notstd/pipe.cpp
@@ -257,7 +257,7 @@ bool NamedPipe::Write(const char *buffer, std::size_t size)
 		return false;
 	mPipeDataWrite.operation = PIPE_OPERATION_WRITE;
 	memcpy(mPipeDataWrite.buffer, buffer, size);
-	if (!::WriteFile(mNamePipe, mPipeDataWrite.buffer, size, NULL,
+	if (!::WriteFile(mNamePipe, mPipeDataWrite.buffer, static_cast<DWORD>(size), NULL,
 		reinterpret_cast<LPOVERLAPPED>(&mPipeDataWrite)))
 	{
 		DWORD lastError = ::GetLastError();
@@ -274,7 +274,8 @@ bool NamedPipe::SyncWrite(const char *buffer, std::size_t size)
 	while (left)
 	{
 		DWORD writed = 0;
-		if (!::WriteFile(mNamePipe, p, left, &writed, NULL) || !writed)
+		if (!::WriteFile(mNamePipe, p, static_cast<DWORD>(left), &writed, NULL)
+			|| !writed)
 			return false;
 		p += writed;
 		left -= writed;
